requirements.txt

streamlit
langchain
openai
langchain-community
langchain
langchain-community
langchain-openai
openai>=1.0
python-dotenv
streamlit
